+++ requests_cache/session.py
@@ -23,7 +23,8 @@ from requests import Session as OriginalSession
 from requests.hooks import dispatch_hook
 from urllib3 import filepost
 
-from .backends import BackendSpecifier, get_valid_kwargs, init_backend
+from . import get_valid_kwargs
+from .backends import BackendSpecifier, init_backend
 from .cache_control import CacheActions, ExpirationTime
 from .cache_keys import normalize_dict
 from .models import AnyResponse, CachedResponse, set_response_defaults